% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData_Mirna.R
\name{GetData_Mirna}
\alias{GetData_Mirna}
\title{GetData_Mirna}
\usage{
GetData_Mirna(Data)
}
\arguments{
\item{Data}{A character should be one of"survival", "GEP", "MF_Target", "MiRNA_Target", "matrix_mirna_go_inter", "matrix_mirna_go_jaccard"}
}
\value{
data
}
\description{
Get the example data
}
