% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGDEscore.R
\name{GetGDEscore}
\alias{GetGDEscore}
\title{GetGDEscore}
\usage{
GetGDEscore(ExpData,Label)
}
\arguments{
\item{ExpData}{A gene expression profile of interest (rows are genes, columns are samples).}

\item{Label}{A character vector consists of "0" and "1" which represent sample class in the gene expression profile. "0" means normal sample and "1" means disease sample.}
}
\value{
A matrix with one column of GDEscore.
}
\description{
Function "GetGDEscore" is used to calculate gene differential expression levels.
}
\examples{
# Obtain the example data
GEP<-GetData_Mirna("GEP")
label<-GetData_Mirna("label")
# Run the function
GDEscore<-GetGDEscore(GEP,label)
}
