% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MirnaData}
\alias{MirnaData}
\title{An environment variable that includes some example data}
\format{
An environment variable
}
\usage{
MirnaData
}
\description{
An environment variable that includes some example data.
matirx_mirna_go_inter:A matrix of Jaccard score between MiRNAs and GOMF.
matirx_mirna_go_jaccard:A matrix consisting of genes shared by MiRNAs targets and GOMF.
MiRNAScore:a ranked list of strong and weak associations with cancer.
MF_Target:GOMF and corresponding targets.
MiRNA_Target:MiRNAs and corresponding targets.
zscore_BRCA:An example gene expression profile.
label:A vector representing the label of the sample of BRCA, where "1" is the disease sample and "0" is the normal sample.
survival:A dataframe including three columns which are sample, status, and time.
}
\keyword{datasets}
